Implement Special:Newsletter
This page allows users to do various actions with specific newsletters.

Possible actions:
* Special:Newsletter
If no id was specified, the user gets redirected to Special:Newsletters so
that they can look at the existing newsletters' list.

* Special:Newsletter/id
If the id was specified and if the newsletter exists, details
of the newsletter are shown in addition to the action buttons (Delete, Manage,
Subscribe|Unsubscribe). The buttons are only shown for logged-in users and
if they have proper access. If there is no existing newsletter by the
specified id, a standard error page is shown.

* Special:Newsletter/id/subscribe|unsubscribe
Allows logged-in users to subscribe and unsubscribe to/from newsletters.
Non-logged in users are redirected to the login form with a custom error
message. The actual form showed will be switched depending on whether the
current user is subscribed or not (i.e., does not depend on URL).

* Special:Newsletter/id/delete
Allows owner to delete the newsletter. Other users are shown a permissions
error page.

TODO for later:
* Merge functionality of Special:ManageNewsletter with this page.
* Make deletions of newsletters more reliable.
* Check current data before submission on (un)subscribe/delete
form and only show success messages if it actually happened (currently showed
always)
* Make Newsletter methods more reliable and scalable.
* Add tests

Bug: T110191
Bug: T107555
Bug: T110490
Change-Id: Ia1834389a7d651cf3d2175b0b77d48cbdc58e814