Improve subscription functionality on Special:Newsletters
If the user is subscribed to the newsletter, they will now see
a red "Unsubscribe" link and if the user is not subscribed, they
will see a green "Subscribe" link.

When JavaScript is enabled, on clicking on the link, an API request
will be made and the user will be either subscribed or unsubscribed.
While the action is in progress, they will see a "(Un)subscribing..."
message and once the action has been completed, it will change to a
"Subscribe" or "Unsubscribe" message depending on the action and the color
will change. The subscriber count will be incremented or decremented
accordingly and a success message will be shown at the corner. This code
was based upon unwatchedPages.js from core.

When JavaScript is disabled, they will be linked to Special:Newsletter/id/(un)subscribe
where they can subscribe or unsubscribe.

There is a minor glitch where the whole table columns shift when the message changes
but my attempt to fix it (by changing the column widths) doesn't really fix it.
I think it will be easier once T108161 is fixed so left it to fix for future.

Bug: T111711
Change-Id: Id39ce73559cb9893f046094e3e974a33e010c9c7