Add throttle for newsletter creations
Autoconfirmed users are allowed to create newsletters by default. This is a
pretty low user access level and could be abused easily. For instance, a
determined vandal could create lots of newsletters with abusive content
in a short period of time. To reduce the number of such cases, add a rate limit
for newsletter creations. By default, it is limited to 3 newsletter creations per hour.

Change-Id: Ia8da7055340163f4bdb4eb32857fd60952b400b3