Add manage form to Special:Newsletter
Right now this form consists of just a simple textarea field containing
the list of publishers for the newsletters. Users with the appropriate
permissions can add and remove publishers by modifying this list.

As this is not very user-friendly, I'd like to enhance this with JavaScript
but before doing that, the manage API needs to accept usernames instead of uids
and this would mean I'd have to update the to-be-removed-soon Special:ManageNewsletter.
So to avoid that, I plan to remove Special:ManageNewsletter first and then
make the changes to the API needed to add a better form powered by JS.

Bug: T107555
Bug: T110327
Change-Id: Id1be756b11ee873779e4638b6d9f6eabfa6963cf