Use a flag to mark newsletters as deleted
When a user deletes a newsletter, the current implementation deletes multiple
rows across four tables in a single transaction, with a total row-count equal
to the sum of issues and subscribers plus one. Even if we were forfeit
atomicity, perform deletes in batches, and do the work via the job queue, there
would still be a substantial risk of overwhelming the database. What we should
do instead is simply introduce a flag that indicates whether a newsletter is
active or not, and turn that flag off when a user deletes a newsletter. This
eliminates a whole class of issues categorically. We could have many thousands
of newsletters with many thousands of issues and many thousands of subscribers
for each and still have a great deal of room to grow further

Bug: T110490
Change-Id: I4e418c1d6185b004763e6bed6d3e87e0227b3a03