NewsletterDb: Replace manual connection management with DBConnRef
Follows-up 7ad0d474e5b4f9, which adopted dependency injection for
LoadBalancer, but regressed by no longer using DBConnRef and doing
manual connection management.

Motivated by T193565, although unlikely to be the cause of it given
the Newsletter extension isn't on on s7/metawiki, nor does it interact
with centralauth. But, it is one of the very few repos that calls
reuseConnection() which is almost always a mistake, hence removing it
to reduce suspicion and remove from search results.

Bug: T193565
Change-Id: I7cd632616520a0105e344adceb0df53c8406bd41