Remove lots of unused code for impossible false returns from IDatabase
Remove unused code due to a misunderstanding in some of the
methods that IDatabase query methods can return false. This
has not been the case for over a decade, preceeding this extension's
existence (it was most likely the result of a misreading doc blocks
relating to the QUERY_SILENCE_ERRORS flag. The docs have since been
clarified to avoid this misunderstanding).

Some of the NewsletterDb methods actually already expected exceptions
so the handling remains as-is in those cases. In particular, it means
those will most likely not return into fatal errors and might also
not cancel the hook from which it gets called.

Methods like deleteNewsletter and restoreNewsletter, were not
catching DB exceptions. Now that their unused code is removed, and
in their callers, it is now clear that when these fail, they will
cause e.g. onArticleDelete to throw and thus cancel the page deletion
with the DBError surfacing to the UI layer (as opposed to through
the Status object message). This is how it behaved already and seems
fine.

The newsletter-delete-failure message did not appear to add value
and other extensions that make write queries don't all have their
own custom DB-failure messages either. Unless the failure is for
reasons within the responsibility of the extension or user (e.g.
not just intermittent infra issues) it seems a custom message
is not needed.

Note that for most of the methods where I removed unused boolean
handling code, the production code callers in SpecialPage, EditPage,
and API classes were already treating the methods as void, which
further indicates that they never returned false or that it never
mattered to the consuiming code.

Also:
* Remove outdated Assert calls in favour of native type hints.
* Clean up doc comments.
* Improve a few variable names.

Change-Id: I078eb09487bd1c3e0076820a31870d1f4ef7a979