Rename variables containing ParserOutput to $parserOutput
It is difficult for human readers (and code search!) to tell the
difference between instances of ParserOutput and OutputPage, because
they are both typically stored in variables named $out or $output
and contain many identically-named methods.

Rename our instances of ParserOutput to make it easier for future
maintainers to tell them apart.

Followup-To: I3eea9e5a052924ac1eeb450fb6b3e955e0b87677
Change-Id: I4deaac6a21de197e2d3f70856b7a3c97b570d524