Undelete newsletters by ID not by name
There could be multiple deleted newsletters with the same name. When
undeleting, only mark one of them as active, chosen arbitrarily.

It's probably neither practical nor worth it to determine which
newsletter you intended to undelete - the only thing preserved from the
old database anyway is the subscribers, and there are workarounds to get
the other one if you need to. All that matters is that we don't corrupt
the database.

Bug: T340374
Change-Id: Ib7503046e82266a8d459b28526a95767e5dede87